requirements.txt
Django==1.8.2
mock==1.0.1
Pillow==2.9.0
py==1.4.28
pytest==2.7.1
wheel==0.24.0
django-registration-redux==1.2

